refactor: migrate AI billing quota to token-based tracking and fix BotWorkflow status updates

+++ services/WorkflowService/src/WorkflowService/Services/BillingQuotaGuard.cs
@@ -35,7 +35,7 @@ public class BillingQuotaGuard(
         }
     }
 
-    public async Task IncrementUsageAsync(Guid? companyId, string usageType, int amount,
+    public async Task IncrementUsageAsync(Guid? companyId, string usageType, long amount,
         CancellationToken ct = default)
     {
         if (!Enabled || companyId is null)
